Add: main.cpp and Makefile
cgi_envのunitテストを作成するため
