+++ test/unit/cgi_env/Makefile
@@ -1,0 +1,52 @@
+NAME		:=	a.out
+
+SRCS		:=	main.cpp \
+				create_cgi_env.cpp \
+				create_request_meta_variables.cpp
+
+OBJ_DIR		:=	objs
+OBJS		:=	$(patsubst %.cpp, $(OBJ_DIR)/%.o, $(notdir $(SRCS)))
+
+CXX			:=	c++
+CXXFLAGS	:=	-std=c++98 -Wall -Wextra -Werror -MMD -MP -pedantic
+
+DEPS		:=	$(OBJS:.o=.d)
+MKDIR		:=	mkdir -p
+
+#--------------------------------------------
+.PHONY	: all
+all		: $(NAME)
+
+$(NAME): $(OBJS)
+	$(CXX) -o $@ $^
+
+$(OBJ_DIR)/%.o: %.cpp
+	@$(MKDIR) $(dir $@)
+	$(CXX) $(CXXFLAGS) -c $< -o $@
+
+$(OBJ_DIR)/%.o: $(TARGET_DIR)/%.cpp
+	@$(MKDIR) $(dir $@)
+	$(CXX) $(CXXFLAGS) -c $< -o $@
+
+.PHONY	: clean
+clean:
+	$(RM) -r $(OBJ_DIR)
+
+.PHONY	: fclean
+fclean: clean
+	$(RM) $(NAME)
+
+.PHONY	: re
+re: fclean all
+
+#--------------------------------------------
+.PHONY	: run
+run: all
+	@./$(NAME)
+
+.PHONY	: val
+val: all
+	@valgrind ./$(NAME)
+
+#--------------------------------------------
+-include $(DEPS)
