+++ test/unit/cgi_env/create_cgi_env.cpp
@@ -18,13 +18,3 @@ const char **create_cgi_env(const std::map<std::string, std::string> &request_me
 	cgi_env[i] = NULL;
 	return (cgi_env);
 }
-
-int main(void) {
-	std::map<std::string, std::string> request_meta_variables = create_request_meta_variables();
-	const char                       **cgi_env = create_cgi_env(request_meta_variables);
-	for (size_t i = 0; cgi_env[i] != NULL; i++) {
-		std::cout << cgi_env[i] << std::endl;
-	}
-	delete[] cgi_env;
-	return (0);
-}
